learning_unit_proposal create business module
